[A2] Extract watch and report domains from cli.py
Move watch-mode helpers (dataclasses, event handler, live screen) into
cli_watch.py and report output helpers (format dispatch, file writing,
baseline output) into cli_report.py. cli.py retains command wiring and
scan orchestration. CLI behavior, flags, and exit codes unchanged.
